fix(build): move funções puras para fora dos arquivos "use server"
O merge do #404 exportou 5 funções puras síncronas de actions/schema.ts
e actions/field-reviews.ts para testá-las diretamente — mas arquivos
"use server" só podem exportar funções async (regra do Next; cada
export vira endpoint RPC). O next build passou a falhar com "Server
Actions must be async functions" e o deploy do frontend no Fly quebrou
em 3 merges consecutivos (f6c1af2, 62c26c3, 55a5dbf) sem chegar em
produção.

Extrai as etapas puras para módulos em lib/:
- lib/schema-backfill.ts: classifyLogEntries,
  reconstructSnapshotsByVersion, matchResponsesToVersions + tipos e
  helpers do backfill de versão de schema
- lib/auto-review-backlog.ts: computeBacklogRows, diffReviewsToRemove,
  compositeKeySet + row-types do backlog de auto-revisão

As actions importam de lib; os dois testes movem junto para
lib/__tests__/. Nenhuma mudança de comportamento.
