+++ frontend/src/actions/field-reviews.ts
@@ -5,14 +5,18 @@ import { createSupabaseAdmin } from "@/lib/supabase/admin";
 import { getAuthUser, getEffectiveMemberId, requireCoordinator } from "@/lib/auth";
 import { buildLoadMap } from "@/lib/load-balancing";
 import { errorMessage } from "@/lib/utils";
-import { computeDivergentFieldNames } from "@/lib/compare-divergence";
-import { isCodingComplete } from "@/lib/coding-completeness";
 import { canonicalPair } from "@/lib/equivalence";
+import { buildEquivalenceMap, type EquivalenceRow } from "@/lib/compare-queue";
 import {
-  buildEquivalenceMap,
-  type EquivalenceByDocField,
-  type EquivalenceRow,
-} from "@/lib/compare-queue";
+  computeBacklogRows,
+  compositeKeySet,
+  diffReviewsToRemove,
+  type AssignmentRow,
+  type ExistingFieldReviewRow,
+  type FieldReviewRow,
+  type HumanResponseRow,
+  type LlmResponseRow,
+} from "@/lib/auto-review-backlog";
 import { resolveBlindVerdict } from "@/lib/arbitration-order";
 import { verdictRequiresJustification } from "@/lib/auto-review-decided";
 import { formatAnswerTechnical } from "@/lib/format-answer";
@@ -20,7 +24,6 @@ import { syncAutoRevisaoAssignmentStatus } from "@/lib/auto-revisao-sync";
 import { syncArbitragemAssignmentStatus } from "@/lib/arbitragem-sync";
 import { revalidatePath } from "next/cache";
 import type {
-  AnswerFieldHashes,
   PydanticField,
   SelfVerdict,
   ArbitrationVerdict,
@@ -802,45 +805,6 @@ export async function submitFinalVerdicts(
 // numero de respostas.
 type SupabaseAdminClient = ReturnType<typeof createSupabaseAdmin>;
 
-export interface HumanResponseRow {
-  id: string;
-  document_id: string;
-  respondent_id: string;
-  answers: Record<string, unknown>;
-  answer_field_hashes: AnswerFieldHashes;
-}
-
-export interface LlmResponseRow {
-  id: string;
-  document_id: string;
-  answers: Record<string, unknown>;
-  answer_field_hashes: AnswerFieldHashes;
-}
-
-export interface ExistingFieldReviewRow {
-  id: string;
-  document_id: string;
-  field_name: string;
-  self_verdict: string | null;
-}
-
-export interface AssignmentRow {
-  project_id: string;
-  document_id: string;
-  user_id: string;
-  type: "auto_revisao";
-  status: "pendente";
-}
-
-export interface FieldReviewRow {
-  project_id: string;
-  document_id: string;
-  field_name: string;
-  human_response_id: string;
-  llm_response_id: string;
-  self_reviewer_id: string;
-}
-
 interface BacklogInputs {
   fields: PydanticField[];
   humanResponses: HumanResponseRow[];
@@ -907,112 +871,9 @@ async function fetchBacklogInputs(
   };
 }
 
-// Varre as respostas humanas completas e calcula quais divergem do LLM —
-// puro, sem I/O (opera só sobre dados já pré-carregados por fetchBacklogInputs).
-export function computeBacklogRows(
-  projectId: string,
-  humanResponses: HumanResponseRow[],
-  llmByDocId: Map<string, LlmResponseRow>,
-  equivByDoc: EquivalenceByDocField,
-  fields: PydanticField[],
-): { assignmentRows: AssignmentRow[]; fieldReviewRows: FieldReviewRow[]; regenerated: number } {
-  const assignmentRows: AssignmentRow[] = [];
-  const fieldReviewRows: FieldReviewRow[] = [];
-  let regenerated = 0;
-
-  for (const human of humanResponses) {
-    const llm = llmByDocId.get(human.document_id);
-    if (!llm) continue;
-
-    // #174: só arbitrar codificações completas. is_partial é sinal inútil de
-    // completude para o humano (quase sempre false), então o filtro de query
-    // não basta: aqui pulamos respostas humanas cuja codificação não está
-    // completa. Espelha o gate inline de saveResponse (allAnswered) via o
-    // mesmo helper. Sem isto, codificações em andamento eram varridas para a
-    // arbitragem e apareciam como "(vazio)" em diversos campos.
-    //
-    // Staleness-aware: passamos answer_field_hashes porque a avaliação é
-    // RETROATIVA (schema atual vs. codificações antigas). Sem isto, um campo
-    // obrigatório adicionado depois (ex.: `medicamento`) tornaria toda
-    // codificação anterior "incompleta", varrendo arbitragens legítimas.
-    if (!isCodingComplete(fields, human.answers ?? {}, human.answer_field_hashes)) {
-      continue;
-    }
-
-    const divergent = computeDivergentFieldNames(
-      fields,
-      [
-        {
-          id: human.id,
-          answers: human.answers ?? {},
-          answerFieldHashes: human.answer_field_hashes,
-        },
-        {
-          id: llm.id,
-          answers: llm.answers ?? {},
-          answerFieldHashes: llm.answer_field_hashes,
-        },
-      ],
-      equivByDoc.get(human.document_id),
-    );
-    if (divergent.length === 0) continue;
-
-    regenerated++;
-    assignmentRows.push({
-      project_id: projectId,
-      document_id: human.document_id,
-      user_id: human.respondent_id,
-      type: "auto_revisao",
-      status: "pendente",
-    });
-    for (const fieldName of divergent) {
-      fieldReviewRows.push({
-        project_id: projectId,
-        document_id: human.document_id,
-        field_name: fieldName,
-        human_response_id: human.id,
-        llm_response_id: llm.id,
-        self_reviewer_id: human.respondent_id,
-      });
-    }
-  }
-
-  return { assignmentRows, fieldReviewRows, regenerated };
-}
-
-// Compartilhado entre diffReviewsToRemove e removeOrphanAssignments: as duas
-// reconciliam uma coleção recém-computada contra uma existente via chave
-// composta document_id+algo. Puro.
-function compositeKeySet<T>(rows: T[], keyFn: (row: T) => string): Set<string> {
-  return new Set(rows.map(keyFn));
-}
-
-// --- Reconcile: quais field_reviews não deveriam mais existir ---
-// O conjunto correto é o que acabou de ser computado em fieldReviewRows.
-// Linhas pendentes (self_verdict IS NULL) fora desse conjunto sao espurias
-// — tipicamente campos que ficaram "stale" apos edicao de schema — e podem
-// ser apagadas. Linhas ja resolvidas pelo pesquisador sao preservadas. Puro.
-export function diffReviewsToRemove(
-  existingReviews: ExistingFieldReviewRow[],
-  fieldReviewRows: FieldReviewRow[],
-): { idsToDelete: string[]; keptResolved: number } {
-  const correctKeys = compositeKeySet(
-    fieldReviewRows,
-    (r) => `${r.document_id}|${r.field_name}`,
-  );
-
-  const idsToDelete: string[] = [];
-  let keptResolved = 0;
-  for (const fr of existingReviews) {
-    if (correctKeys.has(`${fr.document_id}|${fr.field_name}`)) continue;
-    if (fr.self_verdict == null) {
-      idsToDelete.push(fr.id);
-    } else {
-      keptResolved++;
-    }
-  }
-  return { idsToDelete, keptResolved };
-}
+// As etapas puras (computeBacklogRows, diffReviewsToRemove, compositeKeySet)
+// vivem em @/lib/auto-review-backlog — "use server" só pode exportar funções
+// async (regra do Next), então o que é puro e testável fica fora deste arquivo.
 
 // Remove assignments auto_revisao orfaos: sem nenhum field_review restante
 // para o doc+pesquisador e ainda `pendente`. Assignments ja iniciados ou
